run-ragged/pyproject.toml
[tool.poetry]
name = "run-ragged"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
cloud-sql-python-connector = {extras = ["pg8000"], version = "^1.15.0"}
circuitbreaker = "^2.0.0"
python-multipart = "^0.0.9"
fastapi-limiter = "0.1.6"
google-cloud-aiplatform = "^1.38.0"
google-cloud-logging = "3.11.3"
google-cloud-monitoring = "^2.24.0" 
google-cloud-storage = "^2.14.0"
httpx = "0.27.2"
langchain-cli = "0.0.35"
langchain-community = "0.2.5"
langchain-postgres = "^0.0.6"
langchain-google-vertexai = "1.0.5"
langserve = {extras = ["server"], version = ">=0.0.30"}
pgvector = ">=0.1.6,<0.3.0"
pydantic = "<2"
PyPDF2 = "^3.0.0"
psycopg2-binary = "2.9.10"
prometheus-client = "0.21.1"
python = "^3.11"
uvicorn = "0.34.0"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
